+++ src/potcreator.cpp
@@ -17,6 +17,7 @@
 #include "potcreator/potcreator_config.h"
 #include "potcreator/potcreator_gdscript.h"
 #include "potcreator/potcreator_gdscene.h"
+#include "potcreator/potcreator_json.h"
 #include "potcreator/potcreator_sqlite.h"
 #include "potcreator/potcreator_output.h"
 #include "potcreator/potcreator_gen.h"
@@ -183,6 +184,7 @@ int PotCreator::run(int argc, char** argv)
   pimpl->fetchTranslationsForModule<GDScriptModule>();
   pimpl->fetchTranslationsForModule<GDSceneModule>();
   pimpl->fetchTranslationsForModule<SqliteModule>();
+  pimpl->fetchTranslationsForModule<JsonModule>();
 
   genPot(pimpl->cfg, pimpl->translations);